Moved some class files
