+++ src/main/java/fr/utbm/to52/smarthome/controller/CommandHandlerImpl.java
@@ -12,7 +12,7 @@ import org.eclipse.paho.client.mqttv3.IMqttDeliveryToken;
 import org.eclipse.paho.client.mqttv3.MqttCallback;
 import org.eclipse.paho.client.mqttv3.MqttMessage;
 
-import fr.utbm.to52.smarthome.controller.events.Event;
+import fr.utbm.to52.smarthome.controller.events.core.Event;
 
 /**
  * @author Alexandre Guyon
