+++ src/main/java/fr/utbm/to52/smarthome/controller/Controller.java
@@ -9,7 +9,6 @@ import org.json.JSONArray;
 import fr.utbm.to52.smarthome.controller.events.AbortEvent;
 import fr.utbm.to52.smarthome.controller.events.AddNoteEvent;
 import fr.utbm.to52.smarthome.controller.events.AddRingEvent;
-import fr.utbm.to52.smarthome.controller.events.Event;
 import fr.utbm.to52.smarthome.controller.events.GetLogBookEvent;
 import fr.utbm.to52.smarthome.controller.events.GetNoteEvent;
 import fr.utbm.to52.smarthome.controller.events.LightEvent;
@@ -19,10 +18,11 @@ import fr.utbm.to52.smarthome.controller.events.PresenceEvent;
 import fr.utbm.to52.smarthome.controller.events.QuitEvent;
 import fr.utbm.to52.smarthome.controller.events.RingEvent;
 import fr.utbm.to52.smarthome.controller.events.ShutterEvent;
+import fr.utbm.to52.smarthome.controller.events.core.Event;
 import fr.utbm.to52.smarthome.services.AbstractService;
 import fr.utbm.to52.smarthome.services.Service;
 import fr.utbm.to52.smarthome.services.clock.ClockService;
-import fr.utbm.to52.smarthome.services.com.CmdServer;
+import fr.utbm.to52.smarthome.services.clock.CmdServer;
 import fr.utbm.to52.smarthome.services.com.MQTTService;
 import fr.utbm.to52.smarthome.services.couchdb.CouchdbService;
 import fr.utbm.to52.smarthome.services.mail.GmailService;
