+++ src/main/java/fr/utbm/to52/smarthome/controller/events/AddRingEvent.java
@@ -9,6 +9,7 @@ import org.json.JSONObject;
 import org.lightcouch.CouchDbClient;
 
 import fr.utbm.to52.smarthome.controller.Conf;
+import fr.utbm.to52.smarthome.controller.events.core.AbstractEvent;
 import fr.utbm.to52.smarthome.model.cron.RingCron;
 
 /**
