+++ src/main/java/fr/utbm/to52/smarthome/controller/events/LightEvent.java
@@ -9,6 +9,7 @@ import java.util.TimerTask;
 import org.lightcouch.CouchDbClient;
 
 import fr.utbm.to52.smarthome.controller.Conf;
+import fr.utbm.to52.smarthome.controller.events.core.AbstractEvent;
 import fr.utbm.to52.smarthome.services.com.MQTT;
 
 /**
