+++ src/main/java/fr/utbm/to52/smarthome/controller/events/QuitEvent.java
@@ -6,6 +6,7 @@ package fr.utbm.to52.smarthome.controller.events;
 import org.lightcouch.CouchDbClient;
 
 import fr.utbm.to52.smarthome.controller.Controller;
+import fr.utbm.to52.smarthome.controller.events.core.AbstractEvent;
 
 /**
  * @author Alexandre Guyon
