+++ src/main/java/fr/utbm/to52/smarthome/controller/events/core/AbstractDAOComEvent.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package fr.utbm.to52.smarthome.controller.events;
+package fr.utbm.to52.smarthome.controller.events.core;
 
 import org.json.JSONObject;
 import org.lightcouch.CouchDbClient;
