+++ src/main/resources/design-docs/event/views/listByDate/map.js
@@ -1,5 +1,5 @@
 function(doc) {
-	if(doc.$table == "fr.utbm.to52.smarthome.controller.events.StorableEvent"){
+	if(doc.$table == "fr.utbm.to52.smarthome.controller.events.core.StorableEvent"){
 		date = new Date(doc.date);
 		emit([doc.date,date.getDate() + "/" + (date.getMonth() + 1) + "/" + date.getFullYear()], doc);
 	}
