+++ src/main/resources/design-docs/event/views/listByEventName/map.js
@@ -1,5 +1,5 @@
 function(doc) {
-	if(doc.$table == "fr.utbm.to52.smarthome.controller.events.StorableEvent"){
+	if(doc.$table == "fr.utbm.to52.smarthome.controller.events.core.StorableEvent"){
 		emit([doc.date,doc.eventName], doc);
 	}
 }
